fix(order): fix error lines for merging
